refactor: regex 상수 폴더로 이동
